split crient_js and use import
