+++ .babelrc
@@ -3,7 +3,6 @@
     "@babel/preset-react"
   ],
   "plugins": [
-    "@babel/plugin-transform-runtime",
-    "@babel/plugin-transform-modules-commonjs"
+    "@babel/plugin-transform-runtime"
   ]
 }
