+++ bookshelf_app/readOnly.html
@@ -14,7 +14,7 @@
   <script src="https://unpkg.com/react@18/umd/react.development.js" crossorigin></script>
   <script src="https://unpkg.com/react-dom@18/umd/react-dom.development.js" crossorigin></script>
   <!-- Load our React component. -->
-  <script src="/readOnly.js"></script>
+  <script type="module" src="/readOnly.js"></script>
 </body>
 
 </html>
