+++ bookshelf_app/src/home.js
@@ -1,402 +1,9 @@
 'use strict';
-function checkIsValidISBN(isbn) {
-    if (isbn === null) {
-        return false;
-    }
-    const isbn_str = isbn.toString();
-    //現状古い規格は非対応
-    if (isbn_str.length != 13) {
-        return false;
-    }
-    //現状古い規格は非対応
-    if (isbn_str[0] != '9' && isbn_str[1] != '7') {
-        return false;
-    }
-    const check_digit = parseInt(isbn_str.slice(-1)); // バーコードからチェックディジットを抽出する
-    const barcode_digits = isbn_str.slice(0, -1).split(""); // チェックディジットを除いたバーコードの桁を抽出する
-
-    //チェックデジットと照らし合わせる数字を生成
-    let sum = 0;
-    for (let i = 0; i < barcode_digits.length; i++) {
-        if (i % 2 === 0) {
-            sum += parseInt(barcode_digits[i]); // 奇数桁を足す
-        } else {
-            sum += 3 * parseInt(barcode_digits[i]); // 偶数桁を3倍する
-        }
-    }
-
-    //チェックデジットと照らし合わせる
-    if ((sum + check_digit) % 10 === 0) {
-        return true;
-    } else {
-        console.log("Barcode error: " + sum + check_digit);
-        return false;
-    }
-}
-
-async function getBookJson(isbn) {
-    const url = "https://api.openbd.jp/v1/get?isbn=" + isbn;
-    try {
-        const response = await fetch(url);
-        const data = await response.json();
-        if (data && data[0]) {
-            console.log("call isbn api success");
-            return data[0];
-        } else {
-            return { summary: { title: "本の情報を取得できませんでした。isbn:" + isbn } };
-        }
-    } catch (error) {
-        console.error("Error occurred while fetching book data:", error);
-        return { summary: { title: "本のデータを取得中にエラーが発生しました isbn:" + isbn } };
-    }
-}
-
-function BookButton(props) {
-    return <button onClick={() => props.onClick(book)}>{props.bookButtonText()}</button>;
-}
-
-function ShowBooks({ books, bookButton }) {
-    if (books.length === 0) {
-        return <div>本が登録されていません。</div>;
-    }
-    const readStateText = (read_state) => {
-        switch (read_state) {
-            case 0:
-                return "未読";
-            case 1:
-                return "読み止し";
-            case 2:
-                return "読了";
-        }
-    }
-    const listItems = books.map((book, index) => {
-        if (book.detail !== undefined) {
-            return (
-                <tbody key={book.id}>
-                    <tr className="bookDetail">
-                        <td>{book.detail.summary.title}</td>
-                        <td><img src={book.detail.summary.cover}
-                            alt="book_image" width="100" height=" auto" /></td>
-                        <td>:</td>
-                        <td>{readStateText(book.read_state)}</td>
-                        <td>{bookButton(index)}</td>
-                    </tr>
-                </tbody>
-            );
-        } else {
-            return (
-                <tbody key={book.id}>
-                    <tr className="bookDetails"><td></td></tr>
-                </tbody>
-            );
-        }
-
-    });
-    return (
-        <table>
-            {listItems}
-        </table>
-    );
-}
-
-class BookShowState extends React.Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            index: -1,
-        };
-    }
-    ShowBookDetail(bookDetail) {
-        let bookCollateralDetailHtml;
-        if (bookDetail.onix) {
-            bookCollateralDetailHtml = bookDetail.onix.CollateralDetail.TextContent.map((textContent, index) => (
-                <p key={index}>{textContent.Text}</p>
-            ));
-        }
-        return (<div>
-            <h2>{bookDetail.summary.title}</h2>
-            <img src={bookDetail.summary.cover} alt="book_image" width="300" height=" auto" />
-            <p>{bookDetail.summary.author}</p>
-            <div>{bookCollateralDetailHtml}</div>
-            <hr></hr>
-        </div>);
-    }
-    render() {
-        let { books } = this.props;
-        let bookDetailHtml;
-        const bookButton = (index) => {
-            return (<button onClick={() => this.setState({ index: index })} disabled={books[index].detail.onix === undefined} >詳細を見る</button>);
-        }
-        //詳細を見るボタンが押されたら、その本の詳細を表示　
-        if (this.state.index !== -1) {
-            bookDetailHtml = this.ShowBookDetail(books[this.state.index].detail);
-        }
-        return (
-            <div>
-                <hr></hr>
-                {bookDetailHtml}
-                <ShowBooks
-                    books={books}
-                    bookButton={bookButton}
-                />
-            </div>
-        );
-
-    }
-};
-
-class PlayCamera extends React.Component {
-    constructor(props) {
-        super(props);
-
-        this.state = {
-            currentDeviceIndex: 0,
-            devices: [],
-            stream: null
-        };
-
-        this.videoRef = React.createRef();
-        this.handleCameraChange = this.handleCameraChange.bind(this);
-    }
-
-    async componentDidMount() {
-        try {
-            const devices = await navigator.mediaDevices.enumerateDevices();
-            const videoDevices = devices.filter(device => device.kind === "videoinput");
-            this.setState({ devices: videoDevices });
-
-            if (videoDevices.length > 0) {
-                await this.startCamera(videoDevices[this.state.currentDeviceIndex].deviceId);
-            }
-            this.captureImage();
-        } catch (error) {
-            console.error(error);
-        }
-    }
-
-    async componentWillUnmount() {
-        await this.stopCamera();
-    }
-
-    async startCamera(deviceId) {
-        try {
-            const stream = await navigator.mediaDevices.getUserMedia({ video: { deviceId } });
-            this.videoRef.current.srcObject = stream;
-            await this.videoRef.current.play();
-
-            this.setState({ stream });
-        } catch (error) {
-            console.error(error);
-        }
-    }
-
-    async stopCamera() {
-        if (this.state.stream) {
-            this.state.stream.getTracks().forEach(track => track.stop());
-            this.setState({ stream: null });
-        }
-    }
-
-    async handleCameraChange() {
-        const { currentDeviceIndex, devices } = this.state;
-        const nextDeviceIndex = (currentDeviceIndex + 1) % devices.length;
-
-        await this.stopCamera();
-        await this.startCamera(devices[nextDeviceIndex].deviceId);
-        this.captureImage();
-        this.setState({ currentDeviceIndex: nextDeviceIndex });
-    }
-
-    captureImage = () => {
-        if (
-            !this.videoRef.current ||
-            this.videoRef.current.videoWidth === 0 ||
-            this.videoRef.current.videoHeight === 0
-        ) {
-            return;
-        }
-        const canvas = document.createElement("canvas");
-        canvas.width = this.videoRef.current.videoWidth;
-        canvas.height = this.videoRef.current.videoHeight;
-        const ctx = canvas.getContext("2d");
-        ctx.drawImage(this.videoRef.current, 0, 0, canvas.width, canvas.height);
-        const imageData = ctx.getImageData(0, 0, canvas.width, canvas.height);
-        this.props.onReading(imageData);
-        window.requestAnimationFrame(this.captureImage);
-    }
-
-    render() {
-        return (
-            <div className="barcode-reader">
-                <div
-                    style={{
-                        border: "1px solid black",
-                        width: `${this.props.width}px`,
-                        height: `${this.props.height}px`,
-                        marginBottom: "16px"
-                    }}
-                >
-                    <video
-                        ref={this.videoRef}
-                        style={{
-                            objectFit: "contain",
-                            width: "100%",
-                            height: "100%"
-                        }}
-                    />
-                </div>
-                <button onClick={this.handleCameraChange}>カメラ変更</button>
-            </div>
-        );
-    }
-}
-
-class BarcodeReader extends React.Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            imgSrc: null,
-            readedIsbnValue: null,
-        };
-        this.video = { width: 640, height: 480 };
-        this.videoHTML = <></>;
-        this.detectIsbnCode = this.detectIsbnCode.bind(this);
-        if (typeof BarcodeDetector === "function") {
-            this.videoHTML = <PlayCamera onReading={this.detectIsbnCode} width={this.video.width} height={this.video.height} />;
-        }else{
-            this.videoHTML = <div>このブラウザは、バーコードの読み取りに対応していません</div>;
-        }
-    }
-
-    async detectBarcodes(imageData) {
-        try {
-            let barcodes = [];
-            if (typeof BarcodeDetector === "function") {
-                const barcodeDetector = new BarcodeDetector();
-                barcodes = await barcodeDetector.detect(imageData);
-            }
-            if (barcodes.length > 0) {
-                return barcodes;
-            } else {
-                return null;
-            }
-        } catch (error) {
-            console.error(error);
-            return null;
-        }
-    }
-
-    async detectIsbnCode(imageData) {
-        try {
-            let barcodes = await this.detectBarcodes(imageData);
-            if (barcodes && barcodes.length > 0) {
-                for (const barcode of barcodes) {
-                    if (checkIsValidISBN(barcode.rawValue)) {
-                        this.setState({ readedIsbnValue: barcode.rawValue });
-                    }
-                }
-            }
-            return null;
-        } catch (error) {
-            console.error(error);
-            return null;
-        }
-    }
-
-    render() {
-        const { onReadSuccess } = this.props;
-        if (this.state.readedIsbnValue) {
-            onReadSuccess(this.state.readedIsbnValue);
-        }
-        return (
-            <div>
-                { this.videoHTML}
-            </div >
-        );
-    }
-}
-
-class IsbnInputArea extends React.Component {
-    constructor(props) {
-        super(props);
-
-        this.state = {
-            showBarcodeReader: false,
-        };
-    }
-
-    render() {
-        const { showBarcodeReader } = this.state;
-        const { inputingIsbn, inputOnChange, onBarcodeReadSuccess, submitOnClick } = this.props;
-        return (
-            <div className="form-group">
-                <div>ISBNコード
-                    <input type="text" name="isbn" value={inputingIsbn}
-                        inputMode="numeric"
-                        onChange={inputOnChange} />
-                </div>
-                {showBarcodeReader && <BarcodeReader onReadSuccess={(code) => { onBarcodeReadSuccess(code); this.setState({ showBarcodeReader: false }) }} />}
-                {!showBarcodeReader && <button type="button" onClick={() => this.setState({ showBarcodeReader: true })}>バーコード読み取り</button>}
-                {showBarcodeReader && <button type="button" onClick={() => this.setState({ showBarcodeReader: false })}>読み取り中止</button>}
-                <button type="submit" onClick={submitOnClick}>追加</button>
-            </div>
-        );
-    }
-}
-
-class BookAddState extends React.Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            inputingIsbn: "",
-        };
-    }
-    render() {
-        const { submitOnClick, books } = this.props;
-
-        return (
-            <div>
-                <hr></hr>
-                <IsbnInputArea
-                    inputingIsbn={this.state.inputingIsbn}
-                    submitOnClick={() => { submitOnClick(this.state.inputingIsbn); this.setState({ inputingIsbn: "" }); }}
-                    onBarcodeReadSuccess={(barcode) => { this.setState({ inputingIsbn: barcode }); }}
-                    inputOnChange={(e) => { this.setState({ inputingIsbn: e.target.value.replace(/[^0-9]/g, "") }) }}
-                />
-                <hr></hr>
-                <ShowBooks books={books} bookButton={(id) => ""} />
-            </div>
-        );
-    }
-}
-
-function BookReadingChangeState({ books, changeReadState }) {
-    const bookButton = (index) => {
-        return (<button onClick={() => changeReadState(index)}>読書状態変更</button>);
-    }
-    return (
-        <div>
-            <ShowBooks
-                books={books}
-                bookButton={bookButton}
-            />
-        </div>
-    );
-}
-
-function BookDeleteState({ books, deleteBook }) {
-    const bookButton = (index) => {
-        return (<button onClick={() => deleteBook(index)}>本を削除</button>);
-    }
-    return (
-        <div>
-            <ShowBooks
-                books={books}
-                bookButton={bookButton}
-            />
-        </div>
-    );
-}
+import { getBookJson } from './bookUtil.js';
+import BookShowState from './BookShowState.js';
+import BookAddState from './BookAddState.js';
+import BookReadingChangeState from './BookReadingChangeState.js';
+import BookDeleteState from './BookDeleteState.js';
 
 function ModeSelecter(props) {
     return (
