+++ package.json
@@ -16,10 +16,12 @@
     "babel-preset-react-app": "^3.1.2",
     "bcrypt": "^5.1.0",
     "body-parser": "^1.20.2",
+    "core-js": "^3.30.1",
     "express": "^4.18.2",
     "express-session": "^1.17.3",
     "mysql": "^2.18.1",
-    "mysql2": "^3.2.4"
+    "mysql2": "^3.2.4",
+    "regenerator-runtime": "^0.13.11"
   },
   "keywords": [],
   "description": ""
